+++ demos/nextjs/src/components/token-validation.tsx
@@ -2,7 +2,7 @@ import type { SecretKeyType } from '../types'
 import type { TurnstileServerValidationResponse } from '@marsidev/react-turnstile'
 import { type FC, useEffect, useRef, useState } from 'react'
 import { DEMO_SECRET, secretOptions } from '~/constants'
-import Options from './Options'
+import Options from './options'
 
 interface TokenValidationProps {
 	token: string | undefined
