+++ demos/nextjs/src/components/widget-methods.tsx
